requirements.txt
click>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
ollama>=0.4.0
langchain>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.5.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=8.0.0
